I fixed my stock_data.py file and has been integrated to app.py
